fix double

+++ src/ru/itmo/java/basics/pr2/Calculator.java
@@ -28,14 +28,14 @@ public class Calculator {
     public long mul ( long x, long y ){
         return x * y;
     }
-    public int div ( int x, int y){
-        return (x / y) + (x % y);
+    public double div (int x, int y){
+        return (double)(x / y) + (x % y);
     }
     public double div ( double x, double y){
         return x / y;
     }
-    public long div (long x, long y){
-        return  x/ y;
+    public double div (long x, long y){
+        return  (double)x/ y;
     }
     public void enter () {
         System.out.println("\nСложение:");
